refactored to use one component for Price and Volume chart, replaced div with button inside CarouselCoins component
